requirements.txt
# requrements file
# required PyHDB fork is from rybafish repository: https://github.com/rybafish/PyHDB

# tested/proven qt:  5.15.2

PyQt5
PyYAML
PyHDB
cryptography
sqlparse
